requirements.txt
streamlit==1.31.0
pandas==2.2.0
mysql-connector-python==8.3.0
